stock-api/requirements.txt
psycopg2-binary==2.9.11
fastapi==0.75.0
requests==2.32.5
akshare==1.18.22
redis==7.1.1
lunarcalendar==0.0.9
pandas==3.0.0
lunarcalendar==0.0.9
uvicorn==0.40.0
sqlalchemy==2.0.46
